+++ src/util.rs
@@ -31,14 +31,14 @@ impl Direction {
         max: Option<Point2>,
     ) -> Option<Point2> {
         let (try_x, try_y) = match self {
-            Direction::North => (curr.0 as i64, curr.1 as i64 - multiplier),
-            Direction::East => (curr.0 as i64 + multiplier, curr.1 as i64),
-            Direction::South => (curr.0 as i64, curr.1 as i64 + multiplier),
-            Direction::West => (curr.0 as i64 - multiplier, curr.1 as i64),
+            Direction::North => (curr.0, curr.1 - multiplier),
+            Direction::East => (curr.0 + multiplier, curr.1),
+            Direction::South => (curr.0, curr.1 + multiplier),
+            Direction::West => (curr.0 - multiplier, curr.1),
         };
 
-        if (min.is_none() || try_x >= min.unwrap().0 as i64 && try_y >= min.unwrap().1 as i64)
-            && (max.is_none() || try_x < max.unwrap().0 as i64 && try_y < max.unwrap().1 as i64)
+        if (min.is_none() || try_x >= min.unwrap().0 && try_y >= min.unwrap().1)
+            && (max.is_none() || try_x < max.unwrap().0 && try_y < max.unwrap().1)
         {
             Some((try_x, try_y))
         } else {
